# docs/api/workspace.openapi.yaml
openapi: 3.1.0
info:
  title: Oncall Evidence Workspace HTTP API
  version: M2
  description: Implementation-bound contract. Workspace DTOs snake_case; Alertmanager
    timestamps and migrated legacy archive retain original spelling. Bearer console
    OR oncall_session cookie. Cookie mutations require matching Origin and X-CSRF-Token;
    webhook bearer limited to POST /alert. Missing token configuration fails closed503.
    Ordinary reads120/minute/identity. Protected /mcp and /metrics share authentication.
    No endpoint confirms, silences or remediates alerts. M3 global graph deferred;
    absent from paths. Auth/legacy errors can be smaller {error,...}; detailed uniform
    Error applies to workspace handler errors.
servers:
- url: /
security:
- ConsoleBearer: []
- ConsoleCookie: []
paths:
...
  /api/v1/incidents/{id}/topology:
    get:
      operationId: get_api_v1_incidents_id_topology
      summary: Configured read-only service topology
      description: Filters must equal incident scope. Missing config returns capability=not_configured
        and empty arrays. Snapshot source, validity and provenance are explicit. Potential
        impact is inference, not incident evidence; historical_evidence=false.
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/TopologySnapshot'
                  meta:
                    $ref: '#/components/schemas/Meta'
                required:
                - data
                - meta
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthError'
        '403':
          description: Scope or CSRF rejected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthError'
        '404':
          description: Missing record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Version/idempotency conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '413':
          description: Body budget exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: 'Rate limit (Retry-After: 60)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthError'
        '503':
          description: Dependency unavailable; no synthetic success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '504':
          description: Request timeout
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
        description: ''
      - name: environment
        in: query
        required: false
        schema:
          type: string
        description: ''
      - name: service
        in: query
        required: false
        schema:
          type: string
        description: ''
...
components:
  securitySchemes:
    ConsoleBearer:
      type: http
      scheme: bearer
      description: Console token read from configured environment variable
    WebhookBearer:
      type: http
      scheme: bearer
      description: Separate webhook token, POST /alert only
    ConsoleCookie:
      type: apiKey
      in: cookie
      name: oncall_session
  schemas:
...
    TopologyNode:
      type: object
      properties:
        id:
          type: string
        environment:
          type: string
        namespace:
          type: string
        service:
          type: string
        source_ref:
          type: string
      required:
      - id
      - environment
      - namespace
      - service
      - source_ref
    TopologyEdge:
      type: object
      properties:
        id:
          type: string
        source:
          type: string
        target:
          type: string
        kind:
          type: string
        source_ref:
          type: string
        valid_from:
          type: string
        valid_to:
          anyOf:
          - type: string
          - type: 'null'
      required:
      - id
      - source
      - target
      - kind
      - source_ref
      - valid_from
      - valid_to
    PotentialImpact:
      type: object
      properties:
        service_id:
          type: string
        label:
          type: string
        basis:
          type: string
        source_refs:
          type: array
          items:
            type: string
      required:
      - service_id
      - label
      - basis
      - source_refs
    TopologySnapshot:
      type: object
      properties:
        capability:
          type: string
        version:
          type: string
        content_sha256:
          type: string
        source_ref:
          type: string
        valid_from:
          type: string
        valid_to:
          anyOf:
          - type: string
          - type: 'null'
        as_of:
          type: string
        focus_id:
          type: string
        nodes:
          type: array
          items:
            $ref: '#/components/schemas/TopologyNode'
        edges:
          type: array
          items:
            $ref: '#/components/schemas/TopologyEdge'
        potential_impact:
          type: array
          items:
            $ref: '#/components/schemas/PotentialImpact'
        historical_evidence:
          type: boolean
        validity_state:
          type: string
      required:
      - capability
      - version
      - content_sha256
      - source_ref
      - valid_from
      - valid_to
      - as_of
      - focus_id
      - nodes
      - edges
      - potential_impact
      - historical_evidence
      - validity_state
...
    Meta:
      type: object
      properties:
        request_id:
          type: string
        as_of:
          type: string
          format: date-time
        data_state:
          type: string
        mode:
          type: string
          enum:
          - live
        revision:
          type: integer
        next_cursor:
          anyOf:
          - type: string
          - type: 'null'
        truncated:
          type: boolean
        has_more:
          type: boolean
        next_after_sequence:
          type: integer
      required:
      - request_id
      - as_of
      - data_state
      - mode
      - revision
      - next_cursor
      - truncated
    Error:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
        retryable:
          type: boolean
        request_id:
          type: string
      required:
      - error
      - code
      - retryable
      - request_id
    AuthError:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
      required:
      - error
...